+++ contracts/babylon/src/multitest.rs
@@ -95,7 +95,7 @@ mod instantiation {
 
     #[test]
     fn instantiate_light_client_msg_works() {
-        let initial_headers = crate::contract::tests::initial_headers();
+        let initial_headers = babylon_test_utils::initial_headers();
         let header: babylon_bitcoin::BlockHeader = bitcoin::consensus::encode::deserialize_hex("020000003f99814a36d2a2043b1d4bf61a410f71828eca1decbf56000000000000000000b3762ed278ac44bb953e24262cfeb952d0abe6d3b7f8b74fd24e009b96b6cb965d674655dd1317186436e79d").unwrap();
         let params = btc_light_client::msg::InstantiateMsg {
             network: babylon_bitcoin::chain_params::Network::Testnet,
