+++ contracts/babylon/src/multitest/suite.rs
@@ -139,8 +139,7 @@ impl SuiteBuilder {
                         .unwrap_or(1),
                     notify_cosmos_zone: false,
                     btc_light_client_code_id: Some(btc_light_client_code_id),
-                    btc_light_client_initial_headers:
-                        crate::contract::tests::initial_headers_in_hex(),
+                    btc_light_client_initial_headers: babylon_test_utils::initial_headers_in_hex(),
                     btc_light_client_msg: light_client_msg,
                     btc_staking_code_id: Some(btc_staking_code_id),
                     btc_staking_msg: staking_msg,
