+++ packages/test-utils/Cargo.toml
@@ -12,6 +12,7 @@ publish           = true
 babylon-apis     = { path = "../apis", version = "0.14.0" }
 babylon-proto    = { path = "../proto", version = "0.14.0" }
 babylon-bitcoin  = { path = "../bitcoin", version = "0.14.0" }
+bitcoin          = { workspace = true }
 cosmwasm-std     = { workspace = true }
 hex              = { workspace = true }
 k256             = { workspace = true }
